now we can copy/delete the active game object, fixed a bug causing change Z index not affect the scene im.

+++ mario/src/main/java/components/GridLines.java
@@ -16,21 +16,21 @@ public class GridLines extends Component {
     Vector2f cameraPosition = camera.getPosition();
     Vector2f projectionSize = camera.getProjectionSize();
 
-    int firstX = ((int) (cameraPosition.x / Settings.Grid_Width) - 1) * Settings.Grid_Width;
-    int firstY = ((int) (cameraPosition.y / Settings.Grid_Height) -1) * Settings.Grid_Height;
+    float firstX = ((int) (cameraPosition.x / Settings.Grid_Width) - 1) * Settings.Grid_Width;
+    float firstY = ((int) (cameraPosition.y / Settings.Grid_Height) -1) * Settings.Grid_Height;
 
-    int numVerticalLines = (int) (projectionSize.x * camera.getZoom()) / Settings.Grid_Width + 2;
-    int numHorizontalLines = (int) (projectionSize.y * camera.getZoom()) / Settings.Grid_Height + 2;
+    float numVerticalLines = (int) (projectionSize.x * camera.getZoom()) / Settings.Grid_Width + 2;
+    float numHorizontalLines = (int) (projectionSize.y * camera.getZoom()) / Settings.Grid_Height + 2;
 
-    int height = (int) (projectionSize.y * camera.getZoom()) + Settings.Grid_Height;
-    int width = (int) (projectionSize.x * camera.getZoom()) + Settings.Grid_Width;
+    float height = (int) (projectionSize.y * camera.getZoom()) + Settings.Grid_Height;
+    float width = (int) (projectionSize.x * camera.getZoom()) + Settings.Grid_Width;
 
-    int maxLines = Math.max(numVerticalLines, numHorizontalLines);
+    float maxLines = Math.max(numVerticalLines, numHorizontalLines);
 
     Vector3f color = new Vector3f(0.2f, 0.2f, 0.2f);
     for (int i = 0; i < maxLines; i++) {
-      int x = firstX + (i * Settings.Grid_Width);
-      int y = firstY + (i * Settings.Grid_Height);
+      float x = firstX + (i * Settings.Grid_Width);
+      float y = firstY + (i * Settings.Grid_Height);
 
       if (i < numVerticalLines) {
         DebugDraw.addLine2D(new Vector2f(x, firstY), new Vector2f(x, firstY + height), color);
